管理员控制台

+++ src/main/java/com/haole/bupthotelbackend/model/domain/Room.java
@@ -59,7 +59,7 @@ public class Room implements Serializable {
     /**
      * 
      */
-    private Integer acUsageTime;
+    private BigDecimal acUsageTime;
 
     /**
      *
